refactoring data components flow
